refactor: extract AnnotatedScreenshotExport.swift from IssueScreenshotAnnotationTypes.swift
Splits the `AnnotatedScreenshotExport` value out of
IssueScreenshotAnnotationTypes.swift so that file holds only the
`RenderedIssueScreenshotAsset` rendered-asset value.

Byte-identical move of lines 11-20 (includes the load-bearing doc
comment). Both files import Foundation.

IssueScreenshotAnnotationTypes.swift: 20 → 9 lines. Tests: 667.
